% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i2b2_msg.R
\name{add_body}
\alias{add_body}
\title{Add the body to the message}
\usage{
add_body(msg, service, ..., attrib = NULL)
}
\arguments{
\item{msg}{The XML message to add the body to}

\item{service}{The service to request in the body message}

\item{...}{Optionnaly named tags to add inside the body, with their value}

\item{attrib}{A list of XML attributes to add to the service tag}
}
\value{
The XML message list object
}
\description{
Add the body to the XML message
}
\details{
Add the body to the XML message created by base_msg and passed through add_header
base_msg can be piped into add_header and then into add_body to build a message
}
