% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demodata.R
\name{add_observations}
\alias{add_observations}
\title{Add observations to the CRC cell}
\usage{
add_observations(observations, patient_mapping = "", encounter_mapping = "",
  host = "", admin = "", pass = "")
}
\arguments{
\item{observations}{A dataframe of observation facts}

\item{patient_mapping}{The patient mapping table}

\item{encounter_mapping}{The encounter mapping table}

\item{host}{The host to connect to}

\item{admin}{The admin account for the PostgreSQL database}

\item{pass}{The password for the admin account}
}
\description{
Add observations to the CRC cell
}
\details{
The observations dataframe must contain the following columns:
- encounter_ide: the original encounter ID
- patient_ide: the original patient ID
- start_date: the start date of the encounter, as Date object
- concept_cd: the concept to insert
- provider_id: the provider
- modifier_cd: optionnal modifier for the concept
Other observation fact columns can optionnaly be included, 
such as end_date, valtype_cd, tval_char, nval_num, valueflag_cd, units_cd, etc.
}
