% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{add_ont}
\alias{add_ont}
\title{Add an ontology to metadata}
\usage{
add_ont(name, scheme, host = "", admin = "", pass = "")
}
\arguments{
\item{name}{The name of the new ontology}

\item{scheme}{The scheme to use for this ontology}

\item{host}{The host to connect to}

\item{admin}{The admin account for the PostgreSQL database}

\item{pass}{the password for the admin account}
}
\description{
Add an empty ontology space
}
\details{
Add a new empty table to metadata, with indexes
Add a new scheme to the schemes table
Add the corresponding table_access entry for the new table
}
