% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{add_ontology}
\alias{add_ontology}
\title{Add an ontology to i2b2}
\usage{
add_ontology(host = "127.0.0.1", admin, pass, name, scheme, ont,
  modi = NULL, include_code = T)
}
\arguments{
\item{host}{The host to connect to}

\item{admin}{The admin account for the PostgreSQL database}

\item{pass}{the password for the admin account}

\item{name}{The name of the new ontology}

\item{scheme}{The scheme to use for this ontology}

\item{ont}{The ontology to insert}

\item{modi}{The modifiers to insert}

\item{include_code}{Whether to include the code in the label or not}
}
\description{
Add an ontology into i2b2
Add the ontology table and entries in metadata
Populate the ontology table
Populate the concept and modifier dimensions in demodata
}
