% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demodata.R
\name{add_patients_demodata}
\alias{add_patients_demodata}
\title{Add patients to the CRC cell}
\usage{
add_patients_demodata(host, admin, pass, patients, project)
}
\arguments{
\item{host}{The host to connect to}

\item{admin}{The admin account for the PostgreSQL database}

\item{pass}{The password for the admin account}

\item{patients}{A dataframe of patients}

\item{project}{The project to add the patients to}
}
\value{
A patient mapping dataframe for the patients
}
\description{
Add patients to the CRC cell, generate new encrypted IDs,
}
\details{
The patients dataframe must contain the following columns:
- patient_ide: the original patient ID
- birth_date: as a Date object
- death_date: as a Date object
- gender (F or M)
}
