% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imdata.R
\name{add_patients_imdata}
\alias{add_patients_imdata}
\title{Add patients to the IM cell}
\usage{
add_patients_imdata(host, admin, pass, patients, project)
}
\arguments{
\item{host}{The host to connect to}

\item{admin}{The admin account for the PostgreSQL database}

\item{pass}{The password for the admin account}

\item{patients}{A vector of patients IDs to insert}

\item{project}{The project to add the patients to}
}
\description{
Add patients to the IM cell, generate new encrypted IDs,
and create a project site if needed
}
