% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm.R
\name{add_project}
\alias{add_project}
\title{Add a project}
\usage{
add_project(project_id, project_name, host = "", admin = "", pass = "")
}
\arguments{
\item{project_id}{The desired project id}

\item{project_name}{The desired project name}

\item{host}{Address of the host, defaults to 127.0.0.1}

\item{admin}{Name of the database admin account}

\item{pass}{Password of the database admin account}
}
\description{
Add a project with its id, path and name in the hive
}
\details{
Also create a specific database for this project
}
