% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{add_user}
\alias{add_user}
\title{Add an user}
\usage{
add_user(domain, admin, pass, id, name, email, password,
  url = "http://localhost:9090/i2b2/services/PMService/getServices")
}
\arguments{
\item{domain}{The name of the domain to interact with}

\item{pass}{The password for the user}

\item{id}{The id (no spaces, unique) of the new user}

\item{name}{The full name of the new user}

\item{email}{The email of the new user}

\item{password}{The password for the new user}

\item{url}{The URL of the i2b2 cell to communicate with}

\item{username}{The username to connect with}
}
\value{
The XML return message as an httr::content() object
}
\description{
Add an i2b2 user to the instance
}
