% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{add_user_roles}
\alias{add_user_roles}
\title{Add user roles}
\usage{
add_user_roles(admin, pass, id, project, roles,
  url = "http://127.0.0.1:9090/i2b2/services/PMService/getServices")
}
\arguments{
\item{admin}{The username to connect with}

\item{pass}{The password for the user}

\item{id}{The id (no spaces, unique) of the new user}

\item{project}{The project to add the user role to}

\item{roles}{A character vector of roles to add}

\item{url}{The URL of the i2b2 cell to communicate with}
}
\value{
The XML return message as an httr::content() object
}
\description{
Add roles to an i2b2 user
}
