% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{add_users}
\alias{add_users}
\title{Add multiple users}
\usage{
add_users(admin, pass, users,
  url = "http://127.0.0.1:9090/i2b2/services/PMService/getServices",
  domain = "")
}
\arguments{
\item{admin}{The username to connect with}

\item{pass}{The password for the user}

\item{users}{The dataframe containing the users to add}

\item{url}{The URL of the i2b2 cell to communicate with}

\item{domain}{The domain to act on}
}
\value{
The XML return message as an httr::content() object
}
\description{
Add users in bulk from a dataframe
}
\details{
Add users in bulk from a dataframe containing the following columns:
* id: The unique user id to add for each user
* name: The name for each user
* email: The email address for each user
* password: The password for each user
* role: The role to give the user

Any role can be given from the ones defined
(ADMIN, MANAGER, USER, DATA_PROT, DATA_DEID, DATA_LDS, DATA_AGG, DATA_OBFSC)
ADMIN gives all roles for the project, and ADMIN role to project @
MANAGER gives all roles down from DATA_DEID for the project
USER gives USER and DATA_OBFSC roles for the project
Any DATA_* role gives USER and roles down from the DATA_* role given to the project
}
