% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i2b2_msg.R
\name{base_msg}
\alias{base_msg}
\title{Create base XML message}
\usage{
base_msg()
}
\value{
The base msg list object
}
\description{
Create the base XML message
}
\details{
Creates the base XML message as an R list
}
