% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demodata.R
\name{clear_code_lookup}
\alias{clear_code_lookup}
\title{Clear the default code_lookup}
\usage{
clear_code_lookup(host = "127.0.0.1", admin, pass)
}
\arguments{
\item{host}{The host to connect to}

\item{admin}{The admin account for the PostgreSQL database}

\item{pass}{the password for the admin account}
}
\description{
Clear the default codes in code_lookup
}
