% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demodata.R
\name{clear_default_demodata}
\alias{clear_default_demodata}
\title{Clear the default demodata tables}
\usage{
clear_default_demodata(host = "", admin = "", pass = "")
}
\arguments{
\item{host}{The host to connect to}

\item{admin}{The admin account for the PostgreSQL database}

\item{pass}{the password for the admin account}
}
\description{
Clear the default demodata tables
}
