% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imdata.R
\name{clear_default_imdata}
\alias{clear_default_imdata}
\title{Clear the default imdata tables}
\usage{
clear_default_imdata(host, admin, pass)
}
\arguments{
\item{host}{The host to connect to}

\item{admin}{The admin account for the PostgreSQL database}

\item{pass}{the password for the admin account}
}
\description{
Clear the default imdata tables
}
