% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{clear_default_metadata}
\alias{clear_default_metadata}
\title{Clear the default metadata tables}
\usage{
clear_default_metadata(host = "127.0.0.1", admin, pass)
}
\arguments{
\item{host}{The host to connect to}

\item{admin}{The admin account for the PostgreSQL database}

\item{pass}{the password for the admin account}
}
\description{
Clear the default metadata tables
}
\details{
Drop the birn, custom_meta, i2b2 and icd10_icd9 tables
Delete all schemes and table_access
Insert the 'No scheme' scheme
}
