% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{clear_default_workdata}
\alias{clear_default_workdata}
\title{Clear workspace tables}
\usage{
clear_default_workdata(host = "", admin = "", pass = "")
}
\arguments{
\item{host}{The host to connect to}

\item{admin}{The admin account for the PostgreSQL database}

\item{pass}{the password for the admin account}
}
\description{
Clear the workspace tables and inserts default values
}
