% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demodata.R
\name{clear_demodata}
\alias{clear_demodata}
\title{Clear the demodata tables}
\usage{
clear_demodata(project, host = "", admin = "", pass = "")
}
\arguments{
\item{project}{The name of the project}

\item{host}{The host to connect to}

\item{admin}{The admin account for the PostgreSQL database}

\item{pass}{the password for the admin account}
}
\description{
Clear the demodata tables
}
