% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demodata.R
\name{clear_encounter_dimension}
\alias{clear_encounter_dimension}
\title{Clear the default encounters}
\usage{
clear_encounter_dimension(host = "127.0.0.1", admin, pass)
}
\arguments{
\item{host}{The host to connect to}

\item{admin}{The admin account for the PostgreSQL database}

\item{pass}{the password for the admin account}
}
\description{
Clear the default encounters in visit_dimension
}
