% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demodata.R
\name{clear_encounter_mapping}
\alias{clear_encounter_mapping}
\title{Clear the default encounter mappings}
\usage{
clear_encounter_mapping(host = "127.0.0.1", admin, pass)
}
\arguments{
\item{host}{The host to connect to}

\item{admin}{The admin account for the PostgreSQL database}

\item{pass}{the password for the admin account}
}
\description{
Clear the default mappings in encounter_mapping
}
