% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demodata.R
\name{clear_modifier}
\alias{clear_modifier}
\title{Clear the default modifiers}
\usage{
clear_modifier(host = "localhost", admin, pass)
}
\arguments{
\item{host}{The host to connect to}

\item{admin}{The admin account for the PostgreSQL database}

\item{pass}{the password for the admin account}
}
\description{
Clear the default concepts in concept_dimension
}
