% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imdata.R
\name{clear_mpi_demographics}
\alias{clear_mpi_demographics}
\title{Clear the mpi_demographics table}
\usage{
clear_mpi_demographics(host, admin, pass)
}
\arguments{
\item{host}{The host to connect to}

\item{admin}{The admin account for the PostgreSQL database}

\item{pass}{the password for the admin account}
}
\description{
Clear the default mpi_demographics tables
}
