% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demodata.R
\name{clear_observations}
\alias{clear_observations}
\title{Clear the default observation facts}
\usage{
clear_observations(host = "127.0.0.1", admin, pass)
}
\arguments{
\item{host}{The host to connect to}

\item{admin}{The admin account for the PostgreSQL database}

\item{pass}{the password for the admin account}
}
\description{
Clear the default facts in observation_fact
}
