% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imdata.R
\name{clear_project_patients}
\alias{clear_project_patients}
\title{Clear the project_patients tables}
\usage{
clear_project_patients(host, admin, pass)
}
\arguments{
\item{host}{The host to connect to}

\item{admin}{The admin account for the PostgreSQL database}

\item{pass}{the password for the admin account}
}
\description{
Clear the project_patients tables
}
