% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demodata.R
\name{clear_providers}
\alias{clear_providers}
\title{Clear the default providers}
\usage{
clear_providers(host = "127.0.0.1", admin, pass)
}
\arguments{
\item{host}{The host to connect to}

\item{admin}{The admin account for the PostgreSQL database}

\item{pass}{the password for the admin account}
}
\description{
Clear the default providers in provider_dimension
}
