% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clear_table}
\alias{clear_table}
\title{Clear a database table}
\usage{
clear_table(db, table, host = "", admin = "", pass = "")
}
\arguments{
\item{db}{Name of the database}

\item{table}{Name of the table}

\item{host}{The host to connect to}

\item{admin}{The admin account for the PostgreSQL database}

\item{pass}{the password for the admin account}
}
\description{
Clear a database table
}
