% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure.R
\name{create_admin}
\alias{create_admin}
\title{Create system and database admin accounts}
\usage{
create_admin(admin = "i2b2admin", pass = NULL, pass_length = 8)
}
\arguments{
\item{admin}{Name of the admin account to create}

\item{pass}{An arbitrary password if provided, i2b2admin by default}

\item{pass_length}{The length of the generated password, 8 characters by default}
}
\value{
The generated password
}
\description{
Create system and database admin accounts
}
\details{
Create a system and database admin accounts
The name of the account is i2b2admin by default
The shared password can be provided, or randomly generated (length = 8 characters by default)
The system account belongs in the users and wildfly groups
}
