% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_password}
\alias{create_password}
\title{Generate a random password}
\usage{
create_password(length = 8)
}
\arguments{
\item{length}{Length of the desired password}
}
\value{
A password of length length
}
\description{
Generate a random alphanumeric password of any length
}
\details{
Generate a random alphanumeric ([A-Za-z0-9]) password
}
