% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dbUpsert}
\alias{dbUpsert}
\title{Upserta dataframe into a database table}
\usage{
dbUpsert(df, con, table, PK)
}
\arguments{
\item{df}{Dataframe to upsert into the database}

\item{con}{Database connection}

\item{table}{Table in the database in which to push the dataframe}

\item{PK}{Character vector of the primary key(s)}
}
\description{
Upserta dataframe into a database table
}
