% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demodata.R
\name{delete_concept}
\alias{delete_concept}
\title{Delete concepts}
\usage{
delete_concept(host = "127.0.0.1", admin, pass, scheme)
}
\arguments{
\item{host}{The host to connect to}

\item{admin}{The admin account for the PostgreSQL database}

\item{pass}{the password for the admin account}

\item{scheme}{The scheme to delete from the concepts}
}
\description{
Delete concepts from concept_dimension
}
