% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demodata.R
\name{delete_modifier}
\alias{delete_modifier}
\title{Delete modifiers}
\usage{
delete_modifier(host, admin, pass, scheme)
}
\arguments{
\item{host}{The host to connect to}

\item{admin}{The admin account for the PostgreSQL database}

\item{pass}{the password for the admin account}

\item{scheme}{The scheme to delete from the concepts}
}
\description{
Delete modifiers from modifier_dimension
}
