% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{delete_ontology}
\alias{delete_ontology}
\title{Delete an ontology from i2b2}
\usage{
delete_ontology(scheme, host = "", admin = "", pass = "")
}
\arguments{
\item{scheme}{The scheme to use for this ontology}

\item{host}{The host to connect to}

\item{admin}{The admin account for the PostgreSQL database}

\item{pass}{the password for the admin account}
}
\description{
Delete an existing ontology from metadata
}
\details{
Delete the corresponding table
Delete the scheme in schemes table
Delete the entry in table_acess
}
