% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm.R
\name{delete_project}
\alias{delete_project}
\title{Delete a project}
\usage{
delete_project(project_id, host = "", admin = "", pass = "")
}
\arguments{
\item{project_id}{The project to delete}

\item{host}{Address of the host, defaults to 127.0.0.1}

\item{admin}{Name of the database admin account}

\item{pass}{Password of the database admin account}
}
\description{
Delete a project from the hive
}
