% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{delete_users}
\alias{delete_users}
\title{Delete users}
\usage{
delete_users(host = "127.0.0.1", admin, pass, users)
}
\arguments{
\item{host}{The host to connect to}

\item{admin}{The admin account for the PostgreSQL database}

\item{pass}{The password for the database admin}

\item{users}{A character vector of user ids}
}
\description{
Delete i2b2 users from the instance
}
