% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fresh_install.R
\name{fresh_install}
\alias{fresh_install}
\title{Make a fresh install from the base VM}
\usage{
fresh_install(admin, pass, domain_id, domain_name, project_id, project_name)
}
\arguments{
\item{admin}{The admin account for the PostgreSQL database}

\item{pass}{the password for the admin account}

\item{domain_id}{The desired domain_id}

\item{domain_name}{The desired domain_name}

\item{project_id}{The desired project id}

\item{project_name}{The desired project name}
}
\description{
Make a fresh install from the base VM
}
\details{
Set the permissions,
create the admin account for the system, database, and i2b2
Set the domain and project
Delete the default users
Clear the metadata database, and the *_dimension tables in demodata
}
