% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm.R
\name{get_domain}
\alias{get_domain}
\title{Get the domain}
\usage{
get_domain(host = "", admin = "", pass = "")
}
\arguments{
\item{host}{The host to connect to}

\item{admin}{Name of the database admin account}

\item{pass}{Password of the database admin account}
}
\value{
The domain id and name
}
\description{
Get the domain id of the instance
}
