% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_ont}
\alias{get_ont}
\title{Fetch an ontology}
\usage{
get_ont(ont, host = "", admin = "", pass = "")
}
\arguments{
\item{ont}{The name of the ontology, from the c_table_name column in list_ont()}

\item{host}{The host to connect to}

\item{admin}{The admin account for the PostgreSQL database}

\item{pass}{the password for the admin account}
}
\description{
Fetch an ontology
}
