% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_patients_visits}
\alias{import_patients_visits}
\title{Import patients and their visits}
\usage{
import_patients_visits(patients, project)
}
\arguments{
\item{patients}{A formatted dataframe with correctly named columns}

\item{project}{The project to add the data to}
}
\description{
Import patients and their visits
}
\details{
Import the patient_dimension and visit_dimension death_data
As well as creating the mappings and add visit age observations

Structure for patient dataframe:
- patient_ide : character
- encounter_ide : character
- start_date : Date
- end_date : Date
- rum_start : Date
- rum_end : Date
- birth_date : Date
- death_date : Date
- sex_cd : char, 'M' or 'F'
- provider_id : char, 'STRUCT:xxx'
}
