% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demodata.R
\name{list_concepts}
\alias{list_concepts}
\title{List concepts}
\usage{
list_concepts(scheme, project, host = "", admin = "", pass = "")
}
\arguments{
\item{scheme}{The scheme to get the concepts from}

\item{project}{The name of the project}

\item{host}{The host to connect to}

\item{admin}{The admin account for the PostgreSQL database}

\item{pass}{the password for the admin account}
}
\value{
A list of concepts
}
\description{
List the concepts corresponding to a scheme
}
