% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm.R
\name{list_projects}
\alias{list_projects}
\title{List projects}
\usage{
list_projects(host = "", admin = "", pass = "")
}
\arguments{
\item{host}{Address of the hostdefaults to 127.0.0.1}

\item{admin}{Name of the database admin account}

\item{pass}{Password of the database admin account}
}
\value{
The list of projects
}
\description{
List available projects in the hive
}
