% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{list_schemes}
\alias{list_schemes}
\title{List the available schemes}
\usage{
list_schemes(host = "", admin = "", pass = "")
}
\arguments{
\item{host}{The host to connect to}

\item{admin}{The admin account for the PostgreSQL database}

\item{pass}{the password for the admin account}
}
\description{
List the available schemes
}
