% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{list_user_roles}
\alias{list_user_roles}
\title{List user roles}
\usage{
list_user_roles(user, host = "", admin = "", pass = "")
}
\arguments{
\item{user}{An user id}

\item{host}{The host to connect to}

\item{admin}{The admin account for the PostgreSQL database}

\item{pass}{The password for the database admin}
}
\description{
Delete i2b2 users from the instance
}
