% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{list_users}
\alias{list_users}
\title{List users}
\usage{
list_users(host = "", admin = "", pass = "")
}
\arguments{
\item{host}{The host to connect to}

\item{admin}{The admin account for the PostgreSQL database}

\item{pass}{The password for the database admin}
}
\description{
Delete i2b2 users from the instance
}
