% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{populate_ont}
\alias{populate_ont}
\title{Populate an empty ontology table}
\usage{
populate_ont(host = "localhost", admin, pass, ont, name, scheme)
}
\arguments{
\item{host}{The host to connect to}

\item{admin}{The admin account for the PostgreSQL database}

\item{pass}{the password for the admin account}

\item{ont}{The ontology to insert}

\item{name}{The name of the new ontology}

\item{scheme}{The scheme to use for this ontology}
}
\description{
Populate an empty ontology table
}
\details{
Populate an ontology table
ont is a character vector containing all the leaves of the ontology
with their respective path, in the form
code_level1 label_level1/code_level2 label_level2/.../code_leaf label_leaf
The function rebuilds the folders automatically
}
