% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demodata.R
\name{populate_provider}
\alias{populate_provider}
\title{Populate the provider_dimension}
\usage{
populate_provider(ont, scheme, project, host = "", admin = "", pass = "")
}
\arguments{
\item{ont}{The ontology to insert}

\item{scheme}{The scheme to use for this ontology}

\item{project}{The name of the project}

\item{host}{The host to connect to}

\item{admin}{The admin account for the PostgreSQL database}

\item{pass}{the password for the admin account}
}
\description{
Populate the provider_dimension with new providers
}
\details{
ont is a character vector containing all the leaves of the ontology
with their respective path, in the form
code_level1 label_level1/code_level2 label_level2/.../code_leaf label_leaf
}
