% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demodata.R
\name{rebuild_indexes_demodata}
\alias{rebuild_indexes_demodata}
\title{Rebuild the indexes}
\usage{
rebuild_indexes_demodata(project, host = "", admin = "", pass = "")
}
\arguments{
\item{project}{The name of the project}

\item{host}{The host to connect to}

\item{admin}{The admin account for the PostgreSQL database}

\item{pass}{The password for the admin account}
}
\description{
Rebuild the indexes in i2b2demodata
}
