% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i2b2_msg.R
\name{send_msg}
\alias{send_msg}
\title{Send the message}
\usage{
send_msg(msg, cellurl)
}
\arguments{
\item{msg}{The XML message as an R list}

\item{cellurl}{The URL of the i2b2 cell to communicate with}
}
\value{
The XML return message as an httr::content() object
}
\description{
Send the XML message to an i2b2 cell
}
\details{
Send the XML message built by base_msg %>% add_header %>% add_body
to the specified cellurl
}
