% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{service}
\alias{service}
\title{Manage services}
\usage{
service(service, action = c("start", "stop", "restart"), use_sudo = T)
}
\arguments{
\item{service}{The service to manage (pg for short for postgresql-9.1)}

\item{action}{The action to perform}

\item{use_sudo}{Use sudo if not executed by root account (defaults to T)}
}
\description{
Start/stop/restart services
}
