% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure.R
\name{set_permissions}
\alias{set_permissions}
\title{Set the permissions for the webclient and wildfly folders}
\usage{
set_permissions()
}
\description{
Set the permissions wildfly:wildfly +rwx and setgid to the webclient and wildfly folders
}

