% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm.R
\name{set_project}
\alias{set_project}
\title{Set the project}
\usage{
set_project(host = "localhost", admin, pass, project_id, project_name)
}
\arguments{
\item{host}{Address of the host, defaults to 127.0.0.1}

\item{admin}{Name of the database admin account}

\item{pass}{Password of the database admin account}

\item{project_id}{The desired project id}

\item{project_name}{The desired project name}
}
\description{
Set the project id, project path and project name of the instance
}
\details{
Set the project id, project path and project name in the databases
}
