% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demodata.R
\name{add_encounters}
\alias{add_encounters}
\title{Add encounters to the CRC cell}
\usage{
add_encounters(encounters, project, patient_mapping, host = "", admin = "",
  pass = "")
}
\arguments{
\item{encounters}{A dataframe of patients}

\item{project}{The project to add the patients to}

\item{patient_mapping}{The patient mapping table}

\item{host}{The host to connect to}

\item{admin}{The admin account for the PostgreSQL database}

\item{pass}{The password for the admin account}
}
\value{
An encounter mapping dataframe for the encounters
}
\description{
Add encounters to the CRC cell, generate new encrypted IDs
}
\details{
The encounters dataframe must contain the following columns:
- encounter_ide: the original encounter ID
- patient_ide: the original patient ID
- start_date: the start date of the encounter, as Date object
- end_date: the end date of the encounter, as Date object
- inout: I or O if inpatient or outpatient
}
