% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dbPush}
\alias{dbPush}
\title{Push a dataframe into a database table}
\usage{
dbPush(df, con, table)
}
\arguments{
\item{df}{Dataframe to push into the database}

\item{con}{Database connection}

\item{table}{Table in the database in which to push the dataframe}
}
\description{
Push a dataframe into a database table
}
