% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dbUpdate}
\alias{dbUpdate}
\title{Udpdate a dataframe into a database table}
\usage{
dbUpdate(df, con, table, PK)
}
\arguments{
\item{df}{Dataframe to update into the database}

\item{con}{Database connection}

\item{table}{Table in the database in which to push the dataframe}

\item{PK}{Character vector of the primary key(s)}
}
\description{
Udpdate a dataframe into a database table
}
