% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demodata.R
\name{delete_concept}
\alias{delete_concept}
\title{Delete concepts}
\usage{
delete_concept(scheme, project, host = "", admin = "", pass = "")
}
\arguments{
\item{scheme}{The scheme to delete from the concepts}

\item{project}{The name of the project}

\item{host}{The host to connect to}

\item{admin}{The admin account for the PostgreSQL database}

\item{pass}{the password for the admin account}
}
\description{
Delete concepts from concept_dimension
}
