% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{delete_ont}
\alias{delete_ont}
\title{Delete an ontology from metadata}
\usage{
delete_ont(host, admin, pass, scheme)
}
\arguments{
\item{host}{The host to connect to}

\item{admin}{The admin account for the PostgreSQL database}

\item{pass}{the password for the admin account}

\item{scheme}{The scheme to use for this ontology}
}
\description{
Delete an existing ontology from metadata
}
\details{
Delete the corresponding table
Delete the scheme in schemes table
Delete the entry in table_acess
}
