% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demodata.R
\name{list_concepts}
\alias{list_concepts}
\title{List concepts}
\usage{
list_concepts(scheme, host = "", admin = "", pass = "")
}
\arguments{
\item{scheme}{The scheme to get the concepts from}
}
\value{
A list of concepts
}
\description{
List the concepts corresponding to a scheme
}
