% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demodata.R
\name{populate_concept}
\alias{populate_concept}
\title{Populate the concept_dimension}
\usage{
populate_concept(host, admin, pass, ont, modi, name, scheme)
}
\arguments{
\item{host}{The host to connect to}

\item{admin}{The admin account for the PostgreSQL database}

\item{pass}{the password for the admin account}

\item{ont}{The ontology to insert}

\item{modi}{The modifiers to insert}

\item{name}{The name of the new ontology}

\item{scheme}{The scheme to use for this ontology}
}
\description{
Populate the concept_dimension with new concepts
}
\details{
ont is a character vector containing all the leaves of the ontology
with their respective path, in the form
code_level1 label_level1/code_level2 label_level2/.../code_leaf label_leaf
}
