% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure.R
\name{set_domain}
\alias{set_domain}
\title{Set the domain}
\usage{
set_domain(admin, pass, domain_id, domain_name)
}
\arguments{
\item{admin}{Name of the database admin account}

\item{pass}{Password of the database admin account}

\item{domain_id}{The desired domain_id}

\item{domain_name}{The desired domain_name}
}
\description{
Set the domain id and domain name of the instance
}
\details{
Set the domain id and domain name in the databases
and set the domain name in the webclient'
}
